refactor!: extend color scheme instead of creating preset rules
With this commit, this project isn't a preset plugin anymore.

Instead of creating preset and being independent from other presets
(which also makes it hard to implement with other preset's features),
just return an object to add to UnoCSS' `theme.color` option.
So now it can be used with any preset.